+++ src/main/java/com/zuzihe/slackbot/interacivity/SlackInteractiveService.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.zuzihe.slackbot.message.infra.GeminiService;
-import com.zuzihe.slackbot.slack.infra.SlackWebClient;
+import com.zuzihe.slackbot.slack.http.infra.SlackWebClient;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
