+++ src/main/java/com/zuzihe/slackbot/slack/http/infra/SlackWebClient.java
@@ -1,4 +1,4 @@
-package com.zuzihe.slackbot.slack.infra;
+package com.zuzihe.slackbot.slack.http.infra;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
@@ -7,11 +7,11 @@ import org.springframework.web.reactive.function.client.WebClient;
 import java.util.List;
 import java.util.Map;
 
-import static com.zuzihe.slackbot.slack.util.SlackBlockBuilder.section;
-import static com.zuzihe.slackbot.slack.util.SlackBlockBuilder.button;
-import static com.zuzihe.slackbot.slack.util.SlackBlockBuilder.divider;
-import static com.zuzihe.slackbot.slack.util.SlackBlockBuilder.sectionWithButton;
-import static com.zuzihe.slackbot.slack.util.SlackBlockBuilder.urlButton;
+import static com.zuzihe.slackbot.slack.http.util.SlackBlockBuilder.section;
+import static com.zuzihe.slackbot.slack.http.util.SlackBlockBuilder.button;
+import static com.zuzihe.slackbot.slack.http.util.SlackBlockBuilder.divider;
+import static com.zuzihe.slackbot.slack.http.util.SlackBlockBuilder.sectionWithButton;
+import static com.zuzihe.slackbot.slack.http.util.SlackBlockBuilder.urlButton;
 
 
 @Slf4j
