--- src/main/java/com/zuzihe/slackbot/slack/util/SlackSecretManager.java
@@ -1,11 +1,0 @@
-package com.zuzihe.slackbot.slack.util;
-
-import lombok.Getter;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
-
-@Getter
-@Component
-public class SlackSecretManager {
-
-}
